# openapi.yaml
openapi: 3.0.3
info:
  title: RefOnRecord – Resume Verifier API
  description: >
    API for managing resume projects, work experiences, and verification requests.
    Verifiers can be emailed to confirm or reject employment history.
  version: "1.0.0"
  contact:
    name: PWP Spring 2026 Group
    email: Sillah.Babar@student.oulu.fi
servers:
  - url: https://refonrecord.onrender.com/api
    description: Production
  - url: http://localhost:5000/api
    description: Local development
security: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: UUID
  responses:
    BadRequest:
      description: Invalid request body or validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing, invalid, or expired bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated user is not the owner of this resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Experience:
      type: object
      properties:
        experience_id:
          type: integer
        project_id:
          type: integer
        company_name:
          type: string
        position_title:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
          nullable: true
        description:
          type: string
        verification_status:
          type: string
          enum: [not_requested, pending, verified, rejected]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        links:
          type: object
          properties:
            self:
              type: string
            project:
              type: string
            verification_requests:
              type: string
...
paths:
...
  /users/{user_id}/projects/{project_id}/experiences/:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
      - name: project_id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [Experiences]
      summary: List all experiences in a project (owner only, cached 300s)
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of experiences
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Experience"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Experiences]
      summary: Create a work experience entry (owner only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [company_name, position_title, start_date, description]
              additionalProperties: false
              properties:
                company_name:
                  type: string
                  minLength: 1
                  maxLength: 200
                position_title:
                  type: string
                  minLength: 1
                  maxLength: 200
                start_date:
                  type: string
                  format: date
                end_date:
                  type: string
                  format: date
                  nullable: true
                  description: Must be >= start_date if provided
                description:
                  type: string
                  minLength: 1
      responses:
        "201":
          description: Experience created
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experience"
        "400":
          description: Validation error or end_date before start_date
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{user_id}/projects/{project_id}/experiences/{experience_id}/:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
      - name: project_id
        in: path
        required: true
        schema:
          type: integer
      - name: experience_id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [Experiences]
      summary: Get a single experience (owner only, cached 300s)
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Experience
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experience"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Experiences]
      summary: Update an experience (owner only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              additionalProperties: false
              properties:
                company_name:
                  type: string
                  minLength: 1
                  maxLength: 200
                position_title:
                  type: string
                  minLength: 1
                  maxLength: 200
                start_date:
                  type: string
                  format: date
                end_date:
                  type: string
                  format: date
                  nullable: true
                description:
                  type: string
                  minLength: 1
      responses:
        "200":
          description: Updated experience
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experience"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Experiences]
      summary: Delete an experience (owner only)
      security:
        - BearerAuth: []
      responses:
        "204":
          description: Experience deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Verification Requests ────────────────────────────────────────────────────
...